+++ src/ts/Managers/CampaignRefreshManager.ts
@@ -29,7 +29,7 @@ export class CampaignRefreshManager {
 
         this._campaignManager.onCampaign.subscribe((placementId, campaign) => this.onCampaign(placementId, campaign));
         this._campaignManager.onNoFill.subscribe(placementId => this.onNoFill(placementId));
-        this._campaignManager.onError.subscribe((error, placementIds, rawAdPlan, parsedAdPlan) => this.onError(error, placementIds, rawAdPlan, parsedAdPlan));
+        this._campaignManager.onError.subscribe((error, placementIds, rawAdPlan) => this.onError(error, placementIds, rawAdPlan));
         this._campaignManager.onAdPlanReceived.subscribe(refreshDelay => this.onAdPlanReceived(refreshDelay));
     }
 
@@ -139,7 +139,7 @@ export class CampaignRefreshManager {
         this.handlePlacementState(placementId, PlacementState.NO_FILL);
     }
 
-    private onError(error: WebViewError | Error, placementIds: string[], rawAdPlan?: string, parsedAdPlan?: any) {
+    private onError(error: WebViewError | Error, placementIds: string[], rawAdPlan?: string) {
         this.invalidateCampaigns(this._needsRefill, placementIds);
 
         if(error instanceof Error) {
@@ -148,8 +148,7 @@ export class CampaignRefreshManager {
 
         Diagnostics.trigger('plc_request_failed', {
             error: error,
-            rawAdResponse: rawAdPlan,
-            adResponse: parsedAdPlan
+            rawAdResponse: rawAdPlan
         });
         this._nativeBridge.Sdk.logError(JSON.stringify(error));