+++ src/ts/Views/EndScreen.ts
@@ -22,9 +22,10 @@ export class EndScreen extends View {
     private _gameName: string;
     private _privacy: Privacy;
     private _localization: Localization;
+    private _isSwipeToCloseEnabled: boolean = false;
     private _abGroup: number;
 
-    constructor(nativeBridge: NativeBridge, campaign: Campaign, coppaCompliant: boolean, language: string) {
+    constructor(nativeBridge: NativeBridge, campaign: Campaign, coppaCompliant: boolean, language: string, gameId: string) {
         super(nativeBridge, 'end-screen');
         this._coppaCompliant = coppaCompliant;
         this._localization = new Localization(language, 'endscreen');
@@ -88,11 +89,6 @@ export class EndScreen extends View {
         this._bindings = [
             {
                 event: 'click',
-                listener: (event: Event) => this.onDownloadEvent(event),
-                selector: '.game-background, .btn-download, .game-icon'
-            },
-            {
-                event: 'click',
                 listener: (event: Event) => this.onCloseEvent(event),
                 selector: '.btn-close-region'
             },
@@ -102,6 +98,38 @@ export class EndScreen extends View {
                 selector: '.privacy-button'
             }
         ];
+
+        if(this._abGroup === 10 || this._abGroup === 11) {
+            this._bindings.push({
+                event: 'click',
+                listener: (event: Event) => this.onDownloadEvent(event),
+                selector: '.btn-download'
+            });
+        } else {
+            this._bindings.push({
+                event: 'click',
+                listener: (event: Event) => this.onDownloadEvent(event),
+                selector: '.game-background, .btn-download, .game-icon'
+            });
+        }
+
+        if(gameId === '1300023' || gameId === '1300024') {
+            this._isSwipeToCloseEnabled = true;
+
+            this._bindings.push({
+                event: 'swipe',
+                listener: (event: Event) => this.onCloseEvent(event),
+                selector: '.campaign-container, .game-background, .btn.download'
+            });
+        }
+    }
+
+    public render(): void {
+        super.render();
+
+        if(this._isSwipeToCloseEnabled) {
+            (<HTMLElement>this._container.querySelector('.btn-close-region')).style.display = 'none';
+        }
     }
 
     public show(): void {